tjs,core: use a flat namespace
After giving some other appproaches a try this is the simplest.
