+++ examples/resolve.js
@@ -2,7 +2,7 @@
 //
 
 (async function() {
-    const res = await tjs.dns.getaddrinfo(tjs.args[2]);
+    const res = await tjs.getaddrinfo(tjs.args[2]);
     console.log(JSON.stringify(res, undefined, 2));
 })().catch(e => {
     console.log(e);
