+++ src/js/tjs/index.js
@@ -7,30 +7,23 @@ import { createStdin, createStdout, createStderr } from './stdio.js';
 
 const tjs = Object.create(null);
 const noExport = [
-    'setTimeout',
-    'setInterval',
-    'clearTimeout',
-    'clearInterval',
-    'guessHandleType',
-    'XMLHttpRequest',
+    'TTY',
     'Worker',
-    'signal',
+    'XMLHttpRequest',
+    'clearInterval',
+    'clearTimeout',
+    'evalScript',
+    'guessHandle',
+    'hrtimeMs',
+    'loadScript',
     'random',
-    'args',
-    'versions',
+    'setInterval',
+    'setTimeout',
     'wasm'
 ];
 
-tjs.signal = core.signal;
-
 for (const [key, value] of Object.entries(core)) {
     if (noExport.includes(key)) {
-        continue;
-    }
-
-    // tjs.signal.SIGINT etc.
-    if (key.startsWith('SIG')) {
-        tjs.signal[key] = value;
         continue;
     }
 
@@ -40,6 +33,20 @@ for (const [key, value] of Object.entries(core)) {
 // These values should be immutable.
 tjs.args = Object.freeze(core.args);
 tjs.versions = Object.freeze(core.versions);
+
+// For the REPL.
+Object.defineProperty(tjs, '__evalScript', {
+    enumerable: false,
+    configurable: false,
+    writable: false,
+    value: core.evalScript
+});
+Object.defineProperty(tjs, '__loadScript', {
+    enumerable: false,
+    configurable: false,
+    writable: false,
+    value: core.loadScript
+});
 
 // Stdio.
 Object.defineProperty(tjs, 'stdin', {
