+++ src/os.c
@@ -23,7 +23,261 @@
  */
 
 #include "private.h"
+#include "utils.h"
+#include "wasm.h"
 
+#include <curl/curl.h>
+#include <string.h>
+#include <unistd.h>
+
+
+static JSValue tjs_exit(JSContext *ctx, JSValueConst this_val, int argc, JSValueConst *argv) {
+    int status;
+    if (JS_ToInt32(ctx, &status, argv[0]))
+        status = -1;
+    // Reset TTY state  (if it had changed) before exiting.
+    uv_tty_reset_mode();
+    exit(status);
+    return JS_UNDEFINED;
+}
+
+static JSValue tjs_hrtime_ms(JSContext *ctx, JSValueConst this_val, int argc, JSValueConst *argv) {
+    return JS_NewFloat64(ctx, uv_hrtime() / 1e6);
+}
+
+static JSValue tjs_gettimeofday(JSContext *ctx, JSValueConst this_val, int argc, JSValueConst *argv) {
+    uv_timeval64_t tv;
+    int r = uv_gettimeofday(&tv);
+    if (r != 0)
+        return tjs_throw_errno(ctx, r);
+    return JS_NewInt64(ctx, tv.tv_sec * 1000 + (tv.tv_usec / 1000));
+}
+
+static JSValue tjs_uname(JSContext *ctx, JSValueConst this_val, int argc, JSValueConst *argv) {
+    JSValue obj;
+    int r;
+    uv_utsname_t utsname;
+
+    r = uv_os_uname(&utsname);
+    if (r != 0)
+        return tjs_throw_errno(ctx, r);
+
+    obj = JS_NewObjectProto(ctx, JS_NULL);
+    JS_DefinePropertyValueStr(ctx, obj, "sysname", JS_NewString(ctx, utsname.sysname), JS_PROP_C_W_E);
+    JS_DefinePropertyValueStr(ctx, obj, "release", JS_NewString(ctx, utsname.release), JS_PROP_C_W_E);
+    JS_DefinePropertyValueStr(ctx, obj, "version", JS_NewString(ctx, utsname.version), JS_PROP_C_W_E);
+    JS_DefinePropertyValueStr(ctx, obj, "machine", JS_NewString(ctx, utsname.machine), JS_PROP_C_W_E);
+
+    return obj;
+}
+
+static JSValue tjs_guess_handle(JSContext *ctx, JSValueConst this_val, int argc, JSValueConst *argv) {
+    int fd;
+    if (JS_ToInt32(ctx, &fd, argv[0]))
+        return JS_EXCEPTION;
+
+    switch (uv_guess_handle(fd)) {
+        case UV_TTY:
+            return JS_NewString(ctx, "tty");
+        case UV_NAMED_PIPE:
+            return JS_NewString(ctx, "pipe");
+        case UV_FILE:
+            return JS_NewString(ctx, "file");
+        case UV_TCP:
+            return JS_NewString(ctx, "tcp");
+        case UV_UDP:
+            return JS_NewString(ctx, "udp");
+        default:
+            return JS_NewString(ctx, "unknown");
+    }
+}
+
+static JSValue tjs_environ(JSContext *ctx, JSValueConst this_val, int argc, JSValueConst *argv) {
+    uv_env_item_t *env;
+    int envcount, r;
+
+    r = uv_os_environ(&env, &envcount);
+    if (r != 0)
+        return tjs_throw_errno(ctx, r);
+
+    JSValue obj = JS_NewObjectProto(ctx, JS_NULL);
+
+    for (int i = 0; i < envcount; i++) {
+        JS_DefinePropertyValueStr(ctx, obj, env[i].name, JS_NewString(ctx, env[i].value), JS_PROP_C_W_E);
+    }
+
+    uv_os_free_environ(env, envcount);
+
+    return obj;
+}
+
+static JSValue tjs_getenv(JSContext *ctx, JSValueConst this_val, int argc, JSValueConst *argv) {
+    const char *name = JS_ToCString(ctx, argv[0]);
+    if (!name)
+        return JS_EXCEPTION;
+
+    char buf[1024];
+    size_t size = sizeof(buf);
+    char *dbuf = buf;
+    int r;
+
+    r = uv_os_getenv(name, dbuf, &size);
+    if (r != 0) {
+        if (r != UV_ENOBUFS)
+            return tjs_throw_errno(ctx, r);
+        dbuf = js_malloc(ctx, size);
+        if (!dbuf)
+            return JS_EXCEPTION;
+        r = uv_os_getenv(name, dbuf, &size);
+        if (r != 0) {
+            js_free(ctx, dbuf);
+            return tjs_throw_errno(ctx, r);
+        }
+    }
+
+    JSValue ret = JS_NewStringLen(ctx, dbuf, size);
+
+    if (dbuf != buf)
+        js_free(ctx, dbuf);
+
+    return ret;
+}
+
+static JSValue tjs_setenv(JSContext *ctx, JSValueConst this_val, int argc, JSValueConst *argv) {
+    const char *name = JS_ToCString(ctx, argv[0]);
+    if (!name)
+        return JS_EXCEPTION;
+
+    const char *value = JS_ToCString(ctx, argv[1]);
+    if (!value)
+        return JS_EXCEPTION;
+
+    int r = uv_os_setenv(name, value);
+    if (r != 0)
+        return tjs_throw_errno(ctx, r);
+
+    return JS_UNDEFINED;
+}
+
+static JSValue tjs_unsetenv(JSContext *ctx, JSValueConst this_val, int argc, JSValueConst *argv) {
+    const char *name = JS_ToCString(ctx, argv[0]);
+    if (!name)
+        return JS_EXCEPTION;
+
+    int r = uv_os_unsetenv(name);
+    if (r != 0)
+        return tjs_throw_errno(ctx, r);
+
+    return JS_UNDEFINED;
+}
+
+static JSValue tjs_cwd(JSContext *ctx, JSValueConst this_val, int argc, JSValueConst *argv) {
+    char buf[1024];
+    size_t size = sizeof(buf);
+    char *dbuf = buf;
+    int r;
+
+    r = uv_cwd(dbuf, &size);
+    if (r != 0) {
+        if (r != UV_ENOBUFS)
+            return tjs_throw_errno(ctx, r);
+        dbuf = js_malloc(ctx, size);
+        if (!dbuf)
+            return JS_EXCEPTION;
+        r = uv_cwd(dbuf, &size);
+        if (r != 0) {
+            js_free(ctx, dbuf);
+            return tjs_throw_errno(ctx, r);
+        }
+    }
+
+    JSValue ret = JS_NewStringLen(ctx, dbuf, size);
+
+    if (dbuf != buf)
+        js_free(ctx, dbuf);
+
+    return ret;
+}
+
+static JSValue tjs_homedir(JSContext *ctx, JSValueConst this_val, int argc, JSValueConst *argv) {
+    char buf[1024];
+    size_t size = sizeof(buf);
+    char *dbuf = buf;
+    int r;
+
+    r = uv_os_homedir(dbuf, &size);
+    if (r != 0) {
+        if (r != UV_ENOBUFS)
+            return tjs_throw_errno(ctx, r);
+        dbuf = js_malloc(ctx, size);
+        if (!dbuf)
+            return JS_EXCEPTION;
+        r = uv_os_homedir(dbuf, &size);
+        if (r != 0) {
+            js_free(ctx, dbuf);
+            return tjs_throw_errno(ctx, r);
+        }
+    }
+
+    JSValue ret = JS_NewStringLen(ctx, dbuf, size);
+
+    if (dbuf != buf)
+        js_free(ctx, dbuf);
+
+    return ret;
+}
+
+static JSValue tjs_tmpdir(JSContext *ctx, JSValueConst this_val, int argc, JSValueConst *argv) {
+    char buf[1024];
+    size_t size = sizeof(buf);
+    char *dbuf = buf;
+    int r;
+
+    r = uv_os_tmpdir(dbuf, &size);
+    if (r != 0) {
+        if (r != UV_ENOBUFS)
+            return tjs_throw_errno(ctx, r);
+        dbuf = js_malloc(ctx, size);
+        if (!dbuf)
+            return JS_EXCEPTION;
+        r = uv_os_tmpdir(dbuf, &size);
+        if (r != 0) {
+            js_free(ctx, dbuf);
+            return tjs_throw_errno(ctx, r);
+        }
+    }
+
+    JSValue ret = JS_NewStringLen(ctx, dbuf, size);
+
+    if (dbuf != buf)
+        js_free(ctx, dbuf);
+
+    return ret;
+}
+
+static JSValue tjs_random(JSContext *ctx, JSValueConst this_val, int argc, JSValueConst *argv) {
+    size_t size;
+    uint8_t *buf = JS_GetArrayBuffer(ctx, &size, argv[0]);
+    if (!buf)
+        return JS_EXCEPTION;
+
+    uint64_t off = 0;
+    if (!JS_IsUndefined(argv[1]) && JS_ToIndex(ctx, &off, argv[1]))
+        return JS_EXCEPTION;
+
+    uint64_t len = size;
+    if (!JS_IsUndefined(argv[2]) && JS_ToIndex(ctx, &len, argv[2]))
+        return JS_EXCEPTION;
+
+    if (off + len > size)
+        return JS_ThrowRangeError(ctx, "array buffer overflow");
+
+    int r = uv_random(NULL, NULL, buf + off, len, 0, NULL);
+    if (r != 0)
+        return tjs_throw_errno(ctx, r);
+
+    return JS_UNDEFINED;
+}
 
 static JSValue tjs_cpu_info(JSContext *ctx, JSValueConst this_val, int argc, JSValueConst *argv) {
     uv_cpu_info_t *infos;
@@ -127,13 +381,31 @@ static JSValue tjs_network_interfaces(JSContext *ctx, JSValueConst this_val, int
 }
 
 static const JSCFunctionListEntry tjs_os_funcs[] = {
+    TJS_CONST(AF_INET),
+    TJS_CONST(AF_INET6),
+    TJS_CONST(AF_UNSPEC),
+    TJS_CONST(STDIN_FILENO),
+    TJS_CONST(STDOUT_FILENO),
+    TJS_CONST(STDERR_FILENO),
+    TJS_CFUNC_DEF("exit", 1, tjs_exit),
+    TJS_CFUNC_DEF("hrtimeMs", 0, tjs_hrtime_ms),
+    TJS_CFUNC_DEF("gettimeofday", 0, tjs_gettimeofday),
+    TJS_CFUNC_DEF("uname", 0, tjs_uname),
+    TJS_CFUNC_DEF("guessHandle", 1, tjs_guess_handle),
+    TJS_CFUNC_DEF("getenv", 0, tjs_getenv),
+    TJS_CFUNC_DEF("setenv", 2, tjs_setenv),
+    TJS_CFUNC_DEF("unsetenv", 1, tjs_unsetenv),
+    TJS_CFUNC_DEF("cwd", 0, tjs_cwd),
+    TJS_CFUNC_DEF("homedir", 0, tjs_homedir),
+    TJS_CFUNC_DEF("tmpdir", 0, tjs_tmpdir),
+    TJS_CFUNC_DEF("random", 3, tjs_random),
     TJS_CFUNC_DEF("cpuInfo", 0, tjs_cpu_info),
     TJS_CFUNC_DEF("loadavg", 0, tjs_loadavg),
     TJS_CFUNC_DEF("networkInterfaces", 0, tjs_network_interfaces),
 };
 
 void tjs__mod_os_init(JSContext *ctx, JSValue ns) {
-    JSValue obj = JS_NewObjectProto(ctx, JS_NULL);
-    JS_SetPropertyFunctionList(ctx, obj, tjs_os_funcs, countof(tjs_os_funcs));
-    JS_DefinePropertyValueStr(ctx, ns, "os", obj, JS_PROP_C_W_E);
+    JS_SetPropertyFunctionList(ctx, ns, tjs_os_funcs, countof(tjs_os_funcs));
+    // We want environ to be a getter, not a function.
+    CHECK_EQ(TJS_DefineGetter(ctx, ns, tjs_environ, "environ"), 1);
 }
