+++ tests/advanced/test-jsdom.js
@@ -1,6 +1,6 @@
 import assert from '../assert.js';
 
-if (tjs.platform === 'linux' && tjs.environ()['CI']) {
+if (tjs.platform === 'linux' && tjs.environ['CI']) {
     // Skip test on Linux, it creates a stack overflow on some CIs.
     tjs.exit(0);
 }
