+++ tests/test-eval-as-module.js
@@ -4,7 +4,7 @@ import { path } from '@tjs/std';
 
 (async () => {
     const args = [
-        tjs.exepath(),
+        tjs.exepath,
         path.join(import.meta.dirname, 'helpers', 'log-import-meta.js')
     ];
     const proc = tjs.spawn(args);
