+++ tests/test-fswatch.js
@@ -16,21 +16,21 @@ async function sleep(ms) {
 }
 
 (async () => {
-    const tmpDir = await tjs.fs.mkdtemp('test_dirXXXXXX');
-    const watcher = tjs.fs.watch(tmpDir, watchCb);
+    const tmpDir = await tjs.mkdtemp('test_dirXXXXXX');
+    const watcher = tjs.watch(tmpDir, watchCb);
     await sleep(1000);
-    const f = await tjs.fs.mkstemp(path.join(tmpDir, 'test_fileXXXXXX'));
+    const f = await tjs.mkstemp(path.join(tmpDir, 'test_fileXXXXXX'));
     const p = f.path;
     await sleep(1000);
     await f.write(encoder.encode('hello world'));
     await f.close();
     await sleep(1000);
     const newPath = path.join(tmpDir, 'foo');
-    await tjs.fs.rename(p, newPath);
+    await tjs.rename(p, newPath);
     await sleep(1000);
-    await tjs.fs.unlink(newPath);
+    await tjs.unlink(newPath);
     await sleep(1000);
     watcher.close();
-    await tjs.fs.rmdir(tmpDir);
+    await tjs.rmdir(tmpDir);
     assert.eq(eventCount, 5);
 })();
